pyproject.toml
[project]
name = "agriculture-sus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "optuna>=4.4.0",
    "pandas>=2.3.0",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
    "torch>=2.7.1",
    "tqdm>=4.67.1",
]
